requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0

google-generativeai==0.3.2

pyparsing==2.4.7
httplib2==0.22.0
protobuf==4.25.1

pydantic==2.7.0
python-dotenv==1.0.1
